feat: redis 추가
1. redis 캐싱
2. redis 레플리케이션

+++ src/main/java/com/gio/ordersystem/orders/service/OrderService.java
@@ -5,6 +5,9 @@ import com.gio.ordersystem.exception.ExceptionCode;
 import com.gio.ordersystem.orders.domain.Order;
 import com.gio.ordersystem.orders.repository.OrderRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.CachePut;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,27 +15,34 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-
 public class OrderService {
     private final OrderRepository orderRepository;
 
     @Transactional
+    @CachePut(value = "orders", key = "#result.id")
     public Order create(Order order) {
         return orderRepository.save(order);
     }
 
     @Transactional(readOnly = true)
+    @Cacheable(value = "orders", key = "#id") // 조회 시 캐시 사용
     public Order findById(Long id) {
-        return orderRepository.findById(id).orElseThrow(() -> new CustomException(ExceptionCode.ORDER_NOT_FOUND));
+        return orderRepository.findById(id)
+                .orElseThrow(() -> new CustomException(ExceptionCode.ORDER_NOT_FOUND));
     }
 
     @Transactional(readOnly = true)
     public List<Order> getOrdersAfter(Long lastId, int pageSize) {
+        // 전체 데이터는 캐싱하지 않음 (특정 페이지에 대해 필요 시 캐싱 가능)
         return orderRepository.findNextOrders(lastId, pageSize);
     }
 
     @Transactional
+    @CacheEvict(value = "orders", key = "#id") // 삭제 시 해당 캐시 데이터 제거
     public void deleteById(Long id) {
+        if (!orderRepository.existsById(id)) {
+            throw new CustomException(ExceptionCode.ORDER_NOT_FOUND);
+        }
         orderRepository.deleteById(id);
     }
 }
